Added getUnitProperty to the sorting extractor.
Purpose of ythis function is to retrieve information about sorted units, such as spike or unit locations.
The interface is based on dictionary keys, and information is stored in a dectionary within each class.
Added functaionality to  read info to the HS2 sorting extractor, and to both write and read to the numpy class.