Use memmapped read mode, discard directory access
Now uses memmappped h5py to access the recording data, which previously was read into memory in full. Also added a destructor so the open hdf5 file is properly closed when the extractor object is dicarded.

Code that reads MEArec data out of a directory was removed as it is no
longer used in this file format.